requirements.txt
PyQt5==5.15.9
zhipuai==1.0.7
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.27
mysqlclient==2.2.4
pytest==7.4.4
pytest-cov==4.1.0 